application-specific exceptions & more-better exception handling

+++ src/Controller/Item/SpicyKonpeitoController.php
@@ -2,13 +2,14 @@
 namespace App\Controller\Item;
 
 use App\Entity\Inventory;
+use App\Exceptions\PSPInvalidOperationException;
+use App\Exceptions\PSPNotUnlockedException;
 use App\Functions\PetColorFunctions;
 use App\Repository\DragonRepository;
 use App\Service\ResponseService;
 use App\Service\Squirrel3;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Component\HttpKernel\Exception\UnprocessableEntityHttpException;
 use Symfony\Component\Routing\Annotation\Route;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
@@ -34,14 +35,11 @@ class SpicyKonpeitoController extends AbstractController
 
         $dragon = $dragonRepository->findOneBy([ 'owner' => $user ]);
 
-        if(!$dragon)
-            throw new UnprocessableEntityHttpException('You don\'t know any dragons to give the ' . $inventory->getItem()->getName() . ' to...');
-
-        if(!$dragon->getIsAdult())
-            throw new UnprocessableEntityHttpException('Your fireplace dragon, ' . $dragon->getName() . ', is too young for such an adult snack!');
+        if(!$dragon || !$dragon->getIsAdult())
+            throw new PSPNotUnlockedException('Dragon Den');
 
         if(!$inventory->getSpice() || $inventory->getSpice()->getEffects()->getSpicy() === 0)
-            throw new UnprocessableEntityHttpException($dragon->getName() . ' is excited at first, but takes a sniff, and realizes the Konpeitō hasn\'t been properly spiced! (That Konpeitō\'s gotta\'s be SPICY!)');
+            throw new PSPInvalidOperationException($dragon->getName() . ' is excited at first, but takes a sniff, and realizes the Konpeitō hasn\'t been properly spiced! (That Konpeitō\'s gotta\'s be SPICY!)');
 
         $em->remove($inventory);
